# Accelerate `HashMap` and `HashSet` lookup by using `if` based modulo in loops