`HashMap` uses modulo to map hashes and positions to the capacity.
Modulo is slow, so it was exchanged for `fastmod` in https://github.com/godotengine/godot/pull/62327.
However, `fastmod` is still a bottleneck for `HashMap` lookup. Some uses of it can be accelerated with an `if` pseudo modulo, which is faster than `fastmod`.

This change can accelerate `HashMap` lookup by 40%.

## Explanation

On my machine, `fastmod` uses 2 multiplications (3-10 clock cycles[^2]) and a bitshift (1 clock cycle[^3]). It stands to reason that an `if` based modulus should be faster than `fastmod`.

An `if` is an integer comparison (1 clock cycle[^3]) and a mostly predictable branch (0-2 clock cycles[^1] normally, 12 - 25 clock cycles on modulo case). 

It stands to reason that in most cases, an `if` based modulo will be faster than `fastmod`.

## Benchmarks

Theory is well and good, but benchmarks are better. 
I measured an up to 40% improvement in `HashMap `get`, 11% for `insert`, and 2.5% for `erase`.

<details>
<summary>Test Code</summary>
```c++
	for (int size : { 1, 2, 8, 64, 1024, 4096}) {
		{
			size_t time_ns = 0;
			for (int run = 0; run < 20000000 / size; run++) {
				auto t0 = std::chrono::high_resolution_clock::now();
				HashMap<int64_t, int64_t> dictionary;
				for (int idx = 0; idx < size; idx ++) {
					// Test
					dictionary.insert(idx, idx);
				}
				auto t1 = std::chrono::high_resolution_clock::now();
				time_ns += std::chrono::duration_cast<std::chrono::nanoseconds>(t1 - t0).count();
			}

			std::cout << "insert:" << size << std::endl;
			std::cout << time_ns / 1000 / 1000 << "ms\n";
		}
		{
			size_t time_ns = 0;
			for (int run = 0; run < 20000000 / size; run++) {
				HashMap<int64_t, int64_t> dictionary;
				for (int idx = 0; idx < size; idx ++) {
					// Test
					dictionary.insert(idx, idx);
				}
				auto t0 = std::chrono::high_resolution_clock::now();
				for (int idx = 0; idx < size; idx ++) {
					// Test
					dictionary.erase(idx);
				}
				auto t1 = std::chrono::high_resolution_clock::now();
				time_ns += std::chrono::duration_cast<std::chrono::nanoseconds>(t1 - t0).count();
			}

			std::cout << "erase:" << size << std::endl;
			std::cout << time_ns / 1000 / 1000 << "ms\n";
		}
		{
			size_t time_ns = 0;
			for (int run = 0; run < 20000000 / size; run++) {
				HashMap<int64_t, int64_t> dictionary;
				for (int idx = 0; idx < size; idx ++) {
					// Test
					dictionary.insert(idx, idx);
				}
				size_t total = 0;
				auto t0 = std::chrono::high_resolution_clock::now();
				for (int idx = 0; idx < size; idx ++) {
					// Test
					total += dictionary.get(idx);
				}
				auto t1 = std::chrono::high_resolution_clock::now();
				time_ns += std::chrono::duration_cast<std::chrono::nanoseconds>(t1 - t0).count();

				// Prevent compiling this out.
				volatile_size_t = total;
			}

			std::cout << "get:" << size << std::endl;
			std::cout << time_ns / 1000 / 1000 << "ms\n";
		}
	}
```
</details>

On master:
```
insert:1
912ms
erase:1
522ms
get:1
330ms
insert:2
552ms
erase:2
383ms
get:2
174ms
insert:8
310ms
erase:8
376ms
get:8
86ms
insert:64
406ms
erase:64
462ms
get:64
98ms
insert:1024
511ms
erase:1024
587ms
get:1024
105ms
insert:4096
861ms
erase:4096
643ms
get:4096
272ms
```

On this PR:
```
insert:1
913ms
erase:1
521ms
get:1
321ms
insert:2
544ms
erase:2
376ms
get:2
164ms
insert:8
292ms
erase:8
371ms
get:8
61ms
insert:64
376ms
erase:64
459ms
get:64
65ms
insert:1024
475ms
erase:1024
565ms
get:1024
64ms
insert:4096
769ms
erase:4096
627ms
get:4096
190ms
```

## Alternatives

#nazarwadim has proposed in https://github.com/godotengine/godot/pull/90082 to switch to `&` for modulus. However, this would require power of 2 capacities. This may not be desirable, because prime based capacities decrease collision counts on mediocre hashes. We might still consider this, but it should be tested against an optimized prime based version (which this PR is building towards).

[^1]: https://agner.org/optimize/optimizing_cpp.pdf p. 43
[^2]: https://agner.org/optimize/optimizing_cpp.pdf p. 149
[^3]: https://agner.org/optimize/optimizing_cpp.pdf p. 30