@DeeJayLSP 2025-05-20
GDScript has a heavy use of HashMap lookups, so I tested with [GaijinEntertainment/godot-das@master/examples/bunnymark](https://github.com/GaijinEntertainment/godot-das/tree/master/examples/bunnymark's?rgh-link-date=2025-04-25T13%3A42%3A24Z)

At 1000 bunnies (because 10000 is very unlikely and in most cases deceiving with a result that only comes with that much):
| Branch      | FPS  |
|-------------|------|
| Reverted PR | 1262 |
| PR          | 1290 |