@lawnjelly 2025-05-27
`if` branches can be dodgy as _sometimes_ going the other way to branchless code can be faster. But it depends on hardware and the access patterns. In an ideal world we would check on non-desktop, e.g. low end mobile (possibly even web, just in case?).

In this case I asked grok :grin: and it reckons it should be faster on desktop and mobile.

BTW I do wonder if there will be branch prediction, if this is not called in a tight loop (I don't know the access patterns).