@Ivorforce 2025-05-27
> `if` branches can be dodgy as _sometimes_ going the other way to branchless code can be faster. But it depends on hardware and the access patterns. In an ideal world we would check on non-desktop, e.g. low end mobile (possibly even web, just in case?).

Yes, this definitely feels like a micro optimization - it's hard to intuit that the new code is faster than the old one. I would probably not have proposed this change if it didn't lead to such major performance differences for lookup.
On the other hand, branch prediction been ubiquitously important for CPUs for decades, so we can probably expect `if` to stay extremely fast on consumer hardware.

> BTW I do wonder if there will be branch prediction, if this is not called in a tight loop (I don't know the access patterns, and I'm no expert at branch prediction).

Yes, branch prediction happens on every branch (even including short-circuiting `&&` and `||`). As far as I've seen, you can normally assume that the branch that is entered most often is the one that will be predicted.